Upgraded to Vue3, Vuex -> Pinia, and some other stuff
