--- .prettierignore
@@ -1,5 +1,0 @@
-**/*.md
-**/*.svg
-**/*.ejs
-**/*.html
-package.json
